refactor: AIチャットを別コンポーネントに分離しフラット化
- AIChatSection.tsx を新規作成
- AnalysisResult.tsx から入れ子のチャット部分を削除
- スクロール問題の根本原因である複雑な入れ子構造を解消
